+++ lib/components/decorated_navbar.dart
@@ -1,29 +1,25 @@
 part of "../persistent_bottom_nav_bar_v2.dart";
 
 class DecoratedNavBar extends StatelessWidget {
-  DecoratedNavBar({
+  const DecoratedNavBar({
     required this.child,
     super.key,
     this.decoration = const NavBarDecoration(),
-    ImageFilter? filter,
-    this.opacity = 1,
     this.height = kBottomNavigationBarHeight,
-  }) : filter = filter ?? ImageFilter.blur(sigmaX: 3, sigmaY: 3);
+  });
 
   final NavBarDecoration decoration;
-  final ImageFilter filter;
   final Widget child;
-  final double opacity;
   final double height;
 
   @override
   Widget build(BuildContext context) => Stack(
         children: [
-          if (opacity < 1)
+          if ((decoration.color?.opacity ?? 1) < 1 && decoration.filter != null)
             Positioned.fill(
               child: ClipRect(
                 child: BackdropFilter(
-                  filter: decoration.filter ?? filter,
+                  filter: decoration.filter!,
                   child: Container(
                     color: Colors.transparent,
                   ),
@@ -31,11 +27,7 @@ class DecoratedNavBar extends StatelessWidget {
               ),
             ),
           DecoratedBox(
-            decoration: decoration.copyWith(
-              color: opacity != 1
-                  ? decoration.color?.withOpacity(opacity)
-                  : decoration.color,
-            ),
+            decoration: decoration,
             child: SafeArea(
               top: false,
               right: false,