+++ lib/styles/neumorphic_bottom_navbar.dart
@@ -48,12 +48,9 @@ class NeumorphicBottomNavBar extends StatelessWidget {
     ItemConfig item,
     bool isSelected,
   ) =>
-      item.opacity == 1.0
+      neumorphicProperties.decoration?.color?.opacity == 1.0
           ? NeumorphicContainer(
-              decoration: neumorphicProperties.decoration?.copyWith(
-                color: neumorphicProperties.decoration?.color ??
-                    navBarDecoration.color,
-              ),
+              decoration: neumorphicProperties.decoration,
               bevel: neumorphicProperties.bevel,
               curveType: isSelected
                   ? CurveType.emboss
@@ -65,12 +62,7 @@ class NeumorphicBottomNavBar extends StatelessWidget {
           : Container(
               decoration: BoxDecoration(
                 borderRadius: neumorphicProperties.decoration?.borderRadius,
-                color: getBackgroundColor(
-                  context,
-                  navBarConfig.items,
-                  navBarDecoration.color,
-                  navBarConfig.selectedIndex,
-                ),
+                color: neumorphicProperties.decoration?.color,
               ),
               padding: const EdgeInsets.all(6),
               margin: const EdgeInsets.symmetric(horizontal: 4),
@@ -80,8 +72,6 @@ class NeumorphicBottomNavBar extends StatelessWidget {
   @override
   Widget build(BuildContext context) => DecoratedNavBar(
         decoration: navBarDecoration,
-        filter: navBarConfig.selectedItem.filter,
-        opacity: navBarConfig.selectedItem.opacity,
         height: navBarConfig.navBarHeight,
         child: Row(
           mainAxisAlignment: MainAxisAlignment.spaceEvenly,