+++ lib/styles/style_10_bottom_navbar.dart
@@ -107,10 +107,6 @@ class _Style10BottomNavBarState extends State<Style10BottomNavBar>
     }
     return DecoratedNavBar(
       decoration: widget.navBarDecoration,
-      filter:
-          widget.navBarConfig.items[widget.navBarConfig.selectedIndex].filter,
-      opacity:
-          widget.navBarConfig.items[widget.navBarConfig.selectedIndex].opacity,
       height: widget.navBarConfig.navBarHeight,
       child: Row(
         mainAxisAlignment: MainAxisAlignment.spaceAround,