+++ lib/styles/style_11_bottom_navbar.dart
@@ -110,10 +110,6 @@ class _Style11BottomNavBarState extends State<Style11BottomNavBar>
     }
     return DecoratedNavBar(
       decoration: widget.navBarDecoration,
-      filter:
-          widget.navBarConfig.items[widget.navBarConfig.selectedIndex].filter,
-      opacity:
-          widget.navBarConfig.items[widget.navBarConfig.selectedIndex].opacity,
       height: widget.navBarConfig.navBarHeight,
       child: Row(
         mainAxisAlignment: MainAxisAlignment.spaceAround,