+++ lib/styles/style_13_bottom_navbar.dart
@@ -96,8 +96,6 @@ class Style13BottomNavBar extends StatelessWidget {
             const SizedBox(height: 23),
             DecoratedNavBar(
               decoration: navBarDecoration,
-              filter: navBarConfig.selectedItem.filter,
-              opacity: navBarConfig.selectedItem.opacity,
               height: navBarConfig.navBarHeight,
               child: Row(
                 mainAxisAlignment: MainAxisAlignment.spaceAround,