+++ lib/styles/style_3_bottom_navbar.dart
@@ -41,8 +41,6 @@ class Style3BottomNavBar extends StatelessWidget {
   @override
   Widget build(BuildContext context) => DecoratedNavBar(
         decoration: navBarDecoration,
-        filter: navBarConfig.selectedItem.filter,
-        opacity: navBarConfig.selectedItem.opacity,
         height: navBarConfig.navBarHeight,
         child: Row(
           mainAxisAlignment: MainAxisAlignment.spaceAround,