+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- Use `NavBarOverlap.none()` as the default for `navBarOverlap`
+- Removed ItemConfig.opacity. Use the opacity of NavBarDecoration.color instead
+- Removed ItemConfig.filter. Use NavBarDecoration.filter instead
+- Removed default value of NavBarDecoration.filter to allow no filter at all
+
 ## [5.3.1] - 2024-10-03
 ### Fixed
 - Improve documentation on the historyLength (https://github.com/jb3rndt/PersistentBottomNavBarV2/pull/138)